Make byte code instrumentation only once for HTTPRequest
